+++ examples/quantum1d/4.xxz-heisenberg/main.jl
@@ -31,7 +31,8 @@ md"""
 The ground state can then be found by calling `find_groundstate`.
 """
 
-groundstate, cache, delta = find_groundstate(state, H, VUMPS());
+groundstate, cache, info = find_groundstate(state, H, VUMPS());
+info
 
 md"""
 As you can see, VUMPS struggles to converge.
@@ -39,7 +40,8 @@ On its own, that is already quite curious.
 Maybe we can do better using another algorithm, such as gradient descent.
 """
 
-groundstate, cache, delta = find_groundstate(state, H, GradientGrassmann(; maxiter = 20));
+groundstate, cache, info = find_groundstate(state, H, GradientGrassmann(; maxiter = 20));
+info
 
 md"""
 Convergence is quite slow and even fails after sufficiently many iterations.
@@ -71,7 +73,7 @@ Alternatively, the Hamiltonian can be constructed directly on a two-site unit ce
 
 ## H2 = repeat(H, 2); -- copies the one-site version
 H2 = heisenberg_XXX(ComplexF64, Trivial, InfiniteChain(2); spin = 1 // 2)
-groundstate, envs, delta = find_groundstate(
+groundstate, envs, info = find_groundstate(
     state, H2, VUMPS(; maxiter = 100, tol = 1.0e-12)
 );
 
@@ -80,7 +82,7 @@ We get convergence, but it takes an enormous amount of iterations.
 The reason behind this becomes more obvious at higher bond dimensions:
 """
 
-groundstate, envs, delta = find_groundstate(
+groundstate, envs, info = find_groundstate(
     state, H2, IDMRG2(; trunc = truncrank(50), maxiter = 20, tol = 1.0e-12)
 );
 entanglementplot(groundstate)
@@ -125,4 +127,4 @@ Even though the bond dimension is higher than in the example without symmetry, c
 
 println(dim(V1))
 println(dim(V2))
-groundstate, cache, delta = find_groundstate(state, H2, VUMPS(; maxiter = 400, tol = 1.0e-12));
+groundstate, cache, info = find_groundstate(state, H2, VUMPS(; maxiter = 400, tol = 1.0e-12));